feature: working colour palette
Design for the colour palette; moved the submit button from grid presenter into colour palette presenter. Changed some minor things for better readability of the code.
