+++ src/presenters/colorPalettePresenter.jsx
@@ -22,14 +22,37 @@ function colorPalettePresenter(props) {
   ];
 
   const [colorPalette] = React.useState(colorPaletteArray);
-  const [pickedColor, setPickedColor] = React.useState(
-    props.model.paletteColor
-  );
+  const [pickedColor, setPickedColor] = React.useState(props.model.paletteColor);
+  const [chosenLed, setChosenLED] = React.useState(props.model.chosenLED);
+  const initialTimer = props.model.timer;
+  const [timer, setTimer] = React.useState(initialTimer);
+  const [submit, setSubmit] = React.useState(false);
+  const timeoutId = React.useRef(null);
+  const [timeout] = React.useState(15);
+
+  const countTimer = React.useCallback(() => {
+    if (timer <= 0) {
+      setTimer(timeout);
+      setSubmit(false);
+    } else {
+      if (submit) {
+        setTimer(timer - 1);
+        props.model.updateTimer(timer);
+      }
+    }
+  }, [timer, submit]);
+  
+  React.useEffect(() => {
+    timeoutId.current = window.setTimeout(countTimer, 1000);
+    // cleanup function
+    return () => window.clearTimeout(timeoutId.current);
+  }, [timer, countTimer]);
 
   React.useEffect(wasCreatedACB, []);
-
+  
   function observerACB() {
     setPickedColor(props.model.paletteColor);
+    setChosenLED(props.model.chosenLED);
   }
 
   function wasCreatedACB() {
@@ -40,15 +63,25 @@ function colorPalettePresenter(props) {
     return isTakenDownACB;
   }
 
-  function chosenColor(color) {
+  function selectColor(color) {
     props.model.setPaletteColor(color);
+  }
+
+  function updateColor(color, ledNumber) {
+    props.model.updateColorInDatabase(color, ledNumber);
+    setSubmit(true);
   }
 
   return (
     <ColorPaletteView
       colorPaletteArray={colorPalette}
-      colorPicked={chosenColor}
-      pickedColor={pickedColor}
+      colorPicked={selectColor}
+      updateColor={updateColor}
+      chosenLED={chosenLed}
+      chosenColor={pickedColor}
+      timer={timer}
+      timeout={timeout}
+      isLoggedIn={props.isLoggedIn}
     />
   );
 }
